# inventory/investory-all-service-api-spec.yaml
openapi: 3.1.0
info:
  title: Inventory Service
  description: The Inventory Service provides the query, creation, update and deletion of products.
  version: 0.0.1
servers:
  - url: https://mpk-inventory.azurewebsites.net/ 
paths:
  /products:
    get:
      summary: Return all products
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    put:
      summary: Update an existing product
      operationId: updateProduct
      requestBody:
        description: Update a product in the inventory
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
    post:
      summary: Add a new product
      operationId: addProduct
      requestBody:
        description: Add a new product to the inventory
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true                  
  /products/{productId}:
    get:
      summary: Find product by Id
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: Product Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema: 
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    delete:
      summary: Delete a product from the inventory
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: product Id to Delete
          required: true
          schema:
            type: string
  /products/findByModel:
    get:
      summary: Find products by model
      operationId: findProductsByModel
      parameters:
        - name: model
          in: query
          description: Model to find
          required: false
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Invalid status value
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          examples: ["p1"]
        manufacturer:
          type: string
          examples: ["Apple", "Samsung"]
        model:
          type: string
          examples: ["iPhone", "Galaxy"]
        size:
          type: string
          examples: ["64GB", "128GB"]